Removes hardcoded AppState endpoints
Problem:
The AppState API Gateway default endpoints have been hardcoded in the CLI configuration which puts a direct coupling between the CLI configuration and the AppState infrastructure. If the API infrastructure changes, the CLI will break and need to be hot patched. Furthermore, all previous versions of the CLI that support Amplify Admin will fail to work for the affected region(s).

Solution:
Remove the hardcoded endpoints and replace with deterministic custom domains that do not change.

Testing Done:
yarn setup-dev